+++ src/main/java/gs25/hotel/reservation/management/system/repository/hotel/HotelReservationRepository.java
@@ -1,8 +1,8 @@
-package gs25.hotel.reservation.management.system.repository;
+package gs25.hotel.reservation.management.system.repository.hotel;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-import gs25.hotel.reservation.management.system.entity.HotelReservation;
+import gs25.hotel.reservation.management.system.entity.hotel.HotelReservation;
 import gs25.hotel.reservation.management.system.entity.user.User;
 import lombok.extern.slf4j.Slf4j;
